feat: Add NoteRepository interface for managing note data
This commit adds a new file `NoteRepository.java` that defines the `NoteRepository` interface. This interface extends the `JpaRepository` interface and provides methods for retrieving notes based on user and public ID. It also includes custom queries using the `@Query` annotation. This interface will be used for managing note data in the application.