refactor: Update NoteRepository to include user parameter in findByPublicIdAndUser method
The NoteRepository interface was updated to include a new method findByPublicIdAndUser, which takes the publicId and user as parameters. This method retrieves a Note entity based on the publicId and user, while also ensuring that the note is not deleted. This change improves the functionality of the NoteRepository by allowing more specific queries to be made.