refactor: Update password hashes in V3__Populate_users_table.sql
This commit updates the password hashes in the V3__Populate_users_table.sql migration file to use a stronger hashing algorithm. The previous hashes were using the bcrypt algorithm with a cost factor of 12, but they have been updated to use a cost factor of 14 for improved security. This change enhances the security of the application by using stronger password hashes for user authentication.