feat: Add existsByEmail method to UserRepository
The UserRepository interface has been updated to include a new method, existsByEmail, which checks if a user with the given email exists in the database. This addition enhances the functionality of the application by providing a convenient way to check for the existence of a user based on their email address.