refactor: Add BeanConfig class for configuring email service
A new BeanConfig class has been added to the config package. This class is responsible for configuring the email service in the application. It uses the @Value annotation to read a property from the application configuration and determines whether to use a mock implementation or the actual EmailService class. This change improves the flexibility and testability of the email service.